+++ CHANGES.md
@@ -2,9 +2,11 @@
 ### Improvements
  * Improved performance of Json parser by > 30%.
  * Slightly improved performance of MCFunction parser.
+ * Improved error messages for validation of Json objects with union schemas.
 
 ### Fixes
  * Fixed syntax highlighting issues with escape sequences in
+ * Fixed Json validation not reporting missing or illegal properties when a union schema has multiple object schemas.
  * Fixed the description property in pack.mcmeta to be a [raw Json Text](https://minecraft.wiki/w/Raw_JSON_text_format) instead of a plain string